Convert reset-password component to typescript
